requirements/default.txt
Django==1.8.4
django-countries==3.3
django-phonenumber-field==0.7.2
django-reversion==1.9.3
phonenumbers==7.0.9
